Shortcut: Fix  nativeText and portableText when using sequences
So far, nativeText and portableText only considered the singular key
`sequence` property; however, we nowadays promote `sequences` in the
case where you have multiple possible values (which notably happens
when using StandardKey enum entries that map to more than one entry).

To fix the issue, we pick the first entry of the `sequences` list
if it is non-empty, and otherwise fall back to the existing `sequence`
member. Moreover, we decouple the change signals for the two text
properties from the `sequenceChanged` signal, so that we can also signal
a change when `sequences` changes instead.

Amends 15a07dbd30e3b242a0495eb6a3316411b493fab3

Pick-to: 6.10
Fixes: QTBUG-96350
Change-Id: I49f9b5a17020a8417a95652c759756cc7fdda1b1
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>