Avoid the Shapes gradient texture cache growing without bounds
Nothing was ever removed from the cache, so in case of e.g. an
animated gradient, a large number of texture objects would be created
and not released, and memory would quickly fill up.

Fix by changing the caching implementation with a set maximum number
of cached gradient textures.

This introduces a soft limit on the supported number of simultaneously
displayed different gradients in the application. If exceeded, some
objects will show wrong gradient colors. Although unlikely to be
reached, the limit is documented and configurable by environment
variable.

As a driveby, improve the gradient cache key qHash() implementation a
bit, presumably giving better spread for the QCache.

Fixes: QTBUG-142208
Fixes: QTBUG-136553
Pick-to: 6.11 6.10 6.8 6.5
Change-Id: Ie104f27031572e1c392c0a8ef79d09f4a2ba5a8e
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>