Enable declared TextSelection instances to edit text programmatically
The first qml file in tests/auto/qmltest/textedit seems to determine the
import version: switch to versionless imports on all of them so that
we can use new properties in the new test.

In the autotest, reset the url to planets.txt before each test case,
now that we are loading a different file to test duplication of
multiple blocks.

This reverts commit 295953bfc791ee9a4b0660ee72f54c39e35a3973.

[ChangeLog][QtQuick][TextEdit] In addition to TextEdit.cursorSelection,
you can now create non-visual instances of TextSelection and use them to
modify rich text programmatically.

Change-Id: Ia121def3e033054c7cb117e53af877e2596d4e3a
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>