qmlimportscanner: Optionally add version to output again
This is not needed by our own tooling, and in the case of major version
only imports might lead to funny results.
It is however used by Linux distro tooling, and can't cause issuse with
older QML projects that never used the more advanced import features.
The feature is only enabled when the  add-version flag is passed, as we
don't want to risk confusing our own tooling which might not expect the
value to exist.
Amends 6aa4c83bdbd9a8df05c14ac1aaad2f1d44053ec4

Note that we need to adjust the rootPath.json file, as the new file
causes us to find more modules.

Pick-to: 6.10
Fixes: QTBUG-115170
Change-Id: I20d39dacc87f18fc5bbec93aaec5bdaef0c77c4b
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>