FluentWinUI3 style: Support high contrast theme for Button
On Windows platforms there are several high contrast themes. These
themes follow a different and much lighter palette than the normal
FluentWinUI3 one. They mainly use a combination of 4 colors, as
gotten from the system: COLOR_BTNFACE, COLOR_BTNTEXT, COLOR_HIGHLIGHT,
and COLOR_HIGHLIGHTTEXT.
To support these, when detecting that a high contrast theme is in
use, don't initialize and resolve the style palette but just use the
platorm theme's palette as that is initialized with the correct
system colors.
So far only add support for Button.

Task-number: QTBUG-129088
Pick-to: 6.10
Change-Id: I4b241c2e0a89c6d7f151a31a21d600b9e6a7d7b9
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>