Remove ImHiddenText from password echoMode to enable virtual keyboard
ImHiddenText when using the password echoMode makes virtual keyboard
unaccessible if not tagged as accessibility tool, such as the gnome
keyboard for example.

Remove it to enable the use of virtual keyboard.

This change does not have any effect on the text displayed and does not
impact the security/confidential aspect when entering sensitive data in
a textInput.

Enable users who do not have access to a physical keyboard to still
input senstive data, such as passwords, on their machine.

Characters are shown as bullet dots with or without this patch, disregarding what the ImHiddenText flag might suggest.

Fixes: QTBUG-88367
Change-Id: Ifd2a45e90ae022139a8ca89401b278e35bdc6a03
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>