Fix possible crash when adding invalid glyph index to df cache
There have been reports where an invalid glyph index is being
requested and the application later crashes. Unfortunately, it
has thus far not been reproduced in a minimal example, and the
cause of the invalid glyph itself is not known. However, the
crash itself is related to the fact that we assume all
populated glyphs will be added to the glyph data hash. When
it is not, then trouble appears later when release() is called
for the same glyphs.

To at least protect against this crash, we add an empty glyph
data to the hash for the index for the invalid case. There
will be missing glyphs from the output, but at least it will
not crash.

Pick-to: 6.8 6.10 6.11
Task-number: QTBUG-129176
Task-number: QTBUG-143071
Change-Id: Ia8a0234f03bf01943924e476a814d53b6b8c0e74
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>