QtQml: Only hold CU in ResolvedTypeReference where needed
The only place that actually needs the compilation unit is the object
creator. The object creator will only touch type references we need to
create objects from. For everything else, the property cache and the
QQmlType are good enough.

Therefore, always store property cache and QQmlType in a uniform way. In
turn drop the CU where we can.

Task-number: QTBUG-135286
Change-Id: I1dfac132a1580679b165c2f891a8d9439fa14d1e
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>