Android: QQ4A: Add support for multiple signal arguments in QML
This commit adds support for QML signals with multiple arguments to be
used in the QQ4A context.

In the QtQuickView[Content] layer, this commit adds a new overload for
the connectSignalListener API. This new overload takes a Class<?>[]
parameter instead of the Class<T> parameter of the old API. To support
this, the corresponding QtAndroidQuickViewEmbedding native function has
been replaced with one that takes a jclass array as well. There is a
new implementation for the old single-class-argument API that calls the
new API, instead of reimplementing the logic in two places.

Signal connection and disconnection logic has been moved to
QAndroidViewSignalManager, which is called from the
QtAndroidQuickViewEmbedding namespace native functions.

In QAndroidViewSignalManager, the signal connection now happens by
connecting to the qt_metacall QMetaObject function, which the class has
overridden.

In QAndroidViewSignalManager::qt_metacall, the signal is handled by
detecting the type of signal and parsing the data sent by the signal,
either by reading the property of the sender of the signal that the
signal is linked to, or parsing the void* arguments passed into
qt_metacall.

Task-number: QTBUG-124489
Change-Id: I416bb266b3ffe5c947264d942b1afe3349bca2d7
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>