QtQml: Do not store type references for properties
This allows a type A to have a property of type B while B also has
a property of type A. This is very helpful for list-like or otherwise
linked data structures. For example the BaseConstraint type in the
DeltaBlue benchmark wouldn't have been necessary if this feature had
been available before.

It has the side effect of making it impossible to validate inline
component types before they're instantiated. We don't actually have
the type at the point where a property is declared. However, since
you certainly want to instantiate at some point and qmllint will still
help you find the problem, this is acceptable.

We also have to intercept URLs before we create QQmlType instances now.
This is because now we only have the URLs to identify types. Having
multiple QQmlType instances referring to the same logical type wsd a bad
idea anyway.

[ChangeLog][QtQml] You can now have cyclic type references in QML
documents. A QML type A can have a property of type B while B has a
property of type A. You can still not cyclically inherit or
instantiate types in QML documents, of course

[ChangeLog][QtQml] Your URL interceptors may get called more often now,
for URLs QML neglected to intercept before. This means that more URLs
may be constructed from already-intercepted base URLs. URL interceptors
should check for this condition.

Change-Id: I84753a883c2de98e16d1c3826e7e56e897eeafb8
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>