Add a validation mode for the garbage collector
A new environment variable, "QV4_MM_CROSS_VALIDATE_INCREMENTAL_GC", is
introduced that turns on, when set, additional debug behavior in the GC.

In particular, when the mode is enabled and the GC is running in
incremental mode, after every mark phase, some additional validation will
be performed to catch some amount of issues that could be introduced by
the usage of the incremental garbage collector.

In more details, a snapshot of which object was marked by the
incremental run is stored, the mark phase of the GC is then run a second
time in a stop-the-world manner.
The result of the second phase is then compared to the original, making
sure that supposedly bug-induced discrepancies are logged.

Generally, any discrepancy should indicate something that has gone
wrong in one run or the other, possibly both independently.
Nonetheless, for this mode, we assume that non-incremental runs of the
GC can be considered to have a baseline correctness that can be used as
a verifier for the output of the mark phase of a GC run, and in
particular a run of the incremental mode of the GC.

A new state was added to the GC that is intended to run between the end
of the mark phase and the start of the sweep phase.

The implementation for the execution of the `markReady` state was
modified to traverse to this new state during an incremental run when
the relevant environment variable was set during the creation of a
`MemoryManager` instance.

The new state generally implements the algorithm described above.
In particular, a copy of the black bitmaps of each allocator, which are
the result of the mark phase and are enough to pinpoint which part of
the memory was marked, is stored away.
The relevant state that is hold by the GC, is reset to a state that
allows running the mark phase again.
The mark phase is then run from start to finish and the new state of the
black bitmaps of each allocator is compared to the one produced by the
latest run.

Errors are reported when we find that the incremental run has not marked
an object that was considered alive by the non-incremental run, that is,
that we are going to collect an object that is still alive.
Cases where the new run has found an object, that was considered to be
alive by an incremental-run, to be dead, are ignored.
This is due to the incremental run of a GC sometimes being unable to
directly identify an unreachable object as dead, for example when
allocations are performed at certain points in the incremental run.

The implementation of `Managed::className` was modified by extracting
the formatting part out so that it can be accessed as part of the newly
added error reporting.

Some documentation for the new environment variable with a brief and
generic description of the new mode was added to the "Configuring the
JavaScript Engine" documentation page, where similar GC-related
environment variables are documented.

A test was added to ensure that the specific case of discrepancies that
we are interested into are caught by enabling the validation mode.

To allow for the testing process itself to be performed by the
fictitious introduction of bugs of the class we intend to uncover, we
ensure that the entry of the new state in the relevant `GCState` enum
is positioned as if it was part of the sweep phase.
Normally, the state that performs the verification will need to redrain
the stack to take into account changes that can have occurred between
the last state and the start of the verification state, as otherwise
false positives could be introduced by the partial snapshot of the black
bitmaps.
Nonetheless, a redrain can and should re-observe some objects that could
have already been marked in precedence. When this is so, any object that
is unmarked fictitiously, could be marked back again, preventing the
test, which has to mutate the the state at the boundaries of the
computation, from correctly observing the process.
By ensuring that the validation step is performed as if "it was part of
the sweep phase", it will use, during a normal run, the general redrain
process that is commonly part of the execution loop (as performed by
`transition`), so that the embedding of the redrain in the validation
step itself can be avoided.
The test can then perform the necessary run of the GC without passing by
the normal execution loop, knowing that it controls when allocations are
performed, so that it can introduce the necessary mutations at the
boundaries without the risk of some of them being overwritten.

To simplify the testing process, and in particular to avoid having to
capture `qDebug` output during the test run, test-specific code that
saves an intermediate state that can be used as a witness of the
algorithm working correctly was added to the GC, behind the
"QT_BUILD_INTERNAL" flag.

Fixes: QTBUG-135064
Change-Id: If3f9ef029b51b77aaa5b68f349cbb1b20330be70
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>