FluentWinUI3: Improve ComboBox when using high contrast themes
The ComboBox didn't use the correct colors, for users that were using
high contrast themes on Windows.

It relied on StyleImage's to display the background, which we can't use
under high contrast mode.

Fix by using Rectangle's instead, and change colors where it makes
sense to do so.

Task-number: QTBUG-129088
Pick-to: 6.10
Change-Id: I4a4863002325eb48cd93d2e841f63b62824e2d60
Reviewed-by: Doris Verria <doris.verria@qt.io>