QQuickTableView: use correct model when checking if an item is editable
When TableView needs to resolve if a model item is editable, it needs
to call QAbstractItemModel::flags(modelIndex). But as it stood, it would
call this function on the internal proxy model set by TreeModel, using a
QModelIndex from the application model (tappedIndex).

Since tappedIndex belongs to the application model, this patch
makes sure that we ask the application model, and not the proxy
model, if the given model item is editable.

Fixes: QTBUG-136492
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I0011c46f1b07470ebcb83d93aa3d592e7339345e
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>