Android: QQ4A: Fix synchronization issue with statusChanged signal
Do not allow Qt Event Loop to continue spinning until the
QtQmlStatusChangeListener has been notified about the status change in
QQuickView or 1000ms have passed, whichever comes first.

This timing issue could cause signals to not be received by the user in
cases where QML signals were fired too quickly after the QML view was
done loading, especially if the Android UI thread was under high load.

This was due to the fact that the QtQmlStatusChange event was queued on
the Android UI thread, after which the Qt event loop was allowed to
keep spinning, emitting signals and such. In cases where the Android UI
thread was under load, processing the queued QtQmlStatusChanged event
could take long enough to lose the earliest QML signals. This was
detected on the CI Android emulators during signallistener tests.

Fixes: QTBUG-137025
Change-Id: Ibf04b71a556a7e4a2b5b0338e6ac7876b0228ae1
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>