CMake: Force-build qmltestrunner when cross-compiling
The qmltestrunner binary needs to be available as a CMake target, so
that a $<TARGET_FILE:...> genex can work. KDE projects use such a
genex to run their qml tests.

For Windows on ARM, when it is cross-compiled from an x86 Windows
host, we need the binary to match the target platform and arch, to be
able to run the tests.

To achieve both use cases, change the binary back from an app to
a tool to make it available as a target automatically when
calling find_package(Qt6Qml).

To ensure the binary is also cross-built for the target arch when
cross-compiling, set QT_WILL_BUILD_TOOLS and
QT_WILL_RENAME_TOOL_TARGETS to ON locally in the subdirectory.

Add an opt out not to build the target binary (by skipping the local
QT_WILL_BUILD_TOOLS assignment) if QT_NO_TARGET_QMLTESTRUNNER is set.
This won't affect regular yocto builds that set QT_FORCE_BUILD_TOOLS
to ON, so that will work as usual.

This is a temporary workaround, until we can create some kind of
QtFooAppsConfig.cmake file which would expose the binary as an app
target, rather than a tool target, like we currently do for tools.

Amends 722116b32a5c18dde24d15232edfbecf605a240a

Pick-to: 6.10
Task-number: COIN-1211
Fixes: QTBUG-137005
Change-Id: I9b5a59f5bd30c2e30e5578354515e018383f2926
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>