qmllint: Recognize transparent as a color in quick plugin
Besides the named constants and (A)RGB representations,
QColor::fromString (which is used to parse valid strings) also supports
using "transparent". The linter however did not handle this special case
so far.

This commit adds the missing support. We don't add "transparent" to the
list of known color names to mirror the documentation in
QColor::fromString: "transparent" is not actually a color keyword, but a
separate check.

Amends 53ccd321360a6a88c606210631169741041cf540

Fixes: QTBUG-137054
Change-Id: Ic031144b888d3fa892a87c44b4bffdc4c21bebf8
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>