qmllint: Only recurse into QML scopes when searching inline components
When searching for inline components, we perform a BFS starting from a
type and recursing into its base types and children. We previously
didn't check the scope type of the children. This is wasteful as there
is no point in recursing into scopes such as enum or script scopes.

This also has the side effect of fixing an infinite loop in certain
cases where a grouped property's base type is one of its parents. That
type would then get readded to the list of to-be-visited nodes over and
over. The question of the base type of the grouped property will need to
be dealt with in another commit. Created QTBUG-137327.

Amends bfdf1bf797baa96af1ee800c7d6ef40816ced953

Fixes: QTBUG-137035
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ic56087e2cb5ffd33b4659b411ad8954a4c15e0fa
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>