Qml Language: qqmljs.g restructure Identifiers and Keywords
This commit aims to add some clarity on what is considered
QmlIdentifier, JsIdentifier, IdentifierName and related
keywords.
It adds clear reference points for the ES7, also aims to group
tokens in the corresponding smaller sets.

This commit also adds some comments, aiming to explain the
existence of such definitions like JsIdentifier

One of the motivations behind this commit to make it more
clear and explicit for the maintenance and further addition
of keywords, such as, for example virtual and override, which
are going to be ContextualKeywords from the QML perspective,
while still being totally valid Identifiers from the ECMA
perspective.

Change-Id: I3d335bda609e8977c7c720c62229633821b1da13
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>