QML Language: qqmljs.g update future reserved keywords
Initial implementation was based on the 3rd edition of ECMAScript.
Many things have changed since then and most of keywords were removed
from that section, even starting from the 5th edition of the standard.

This commit removes most of these reserved keywords, while leaving out
some which might potentially be useful for QML in the future.

Change-Id: I0190a6ea6d08b5d0797eb46c6636ffdb295b4565
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>