Make QQmlJSScope ctor private
We generally want scopes to be created via the create factory function,
so that they already end up in a QQmlJSScope::Ptr, as that's what we
normally operate on.

Adjust the few remaining use cases by adding more helper functions
(cloneInto to "copy-assign"; resetForReparse to clear the existing data
from a QQmlJSScope, which is done by assigning a defaulted constructed
instance to it, and then repopulating some fields).

Change-Id: I4eaa4a1e4e6b3f89a636571bd13e8649fc837a41
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>