FluentWinUI3: Improve Slider and RangeSlider with high contrast themes
The Slider and RangeSlider wasn't using the correct colors, for users
that had enabled a high contrast theme. They also relied on StyleImages
for both the backgrounds and the handles.

Use basic Rectangles instead of StyleImages when using a high contrast
theme, with the appropriate colors, in order to make the control look
as similar as possible to the native WinUI 3 controls.

Task-number: QTBUG-129088
Pick-to: 6.10
Change-Id: Ib30a10c15124ab444b9ca4dd501d6e2c75478fcc
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>