Add some missing Q_MANAGED_TYPE
`ArrayData`, `MemberData` and `StringOrSymbol` are the only direct
children of `Managed` that do not define their own type-tag, thus
inheriting `Type_Invalid`.

Both for consistency with other similar elements and as it can be useful
to indiscriminately inspect the type-tag while debugging, a type-tag has
been added through the use of `Q_MANAGED_TYPE` to the relevant types.

Change-Id: Ia13e1d6f0adc5e8e3f8f1d27c4e36a6cb35a56f2
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>