Remove leftovers debug statements in the GC code-base
Some of them don't compile, albeit they could be trivially fixed to.

More generally, they can easily be added in locally and temporarily when
required, while the removal avoids the noise they produce in their
commented-out state, which, supposedly, is going to be their common
state.

Furthermore, the person that generally left them in is not currently
working on the code-base anymore and recent work on this part of the
code-base has been performed by other individuals that haven't found the
commented out statements useful.

Change-Id: If7da1055f7373872fc60e2baf07d2fdd6e029043
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>