Doc: Remove leftovers mention of allocations being black in the GC
Allocations have defaulted to a white state since
bb3999a22d3282911005865135c9d6d49ae65728.

Thus a paragraph that stated that allocations are black during a gc
cycle in the design documentation for the garbage collector was
substituted with a similar paragraph stating that allocations are white.

A phrase that referred to allocations being black as part of the
motivations to avoid a Steel style barrier was rephrased to ensure that
it is clear that allocations are not currently black.

Change-Id: Idf4bcf082076b2295fe69ede164b4e65c5289606
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>