Remove some unnecessary conditional compilation from in the GC code-base
Certain "QT_NO_DEBUG"-based conditional compilation directives are now
removed in favor of always compiling the originally conditional code.

Supposedly, the original intention might have been to avoid additional
calls in the relevant tight loop when they are unnecessary, as
"QT_NO_DEBUG" makes "Q_ASSERT" compile to a discarding usage of the
assert and the intermediate valus that were hidden are only used as part
of the assert.

Nonetheless, the discarding usage of the assert allows to check that the
condition itself would compile, which provides for an easier
maintenance.
Furthermore we would generally expect that, for the "QT_NO_DEBUG" case,
the compiler would simply optimize the unused elements out.

Change-Id: I3b7334763ed08a81e70e0503997783a7e38b38da
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>