QQuickFlickable: fix jump back when scrolling to edge
Since de591d440c2f43a7e94ef4c93db757d254fa7947, we use [vMoved] to check
if it is the first move in dragging. However [vMoved] can be reset to
false during dragging, and this causes the content to jump back when
scrolling to the edge of the flickable.
(for example when pixel delta is zero, see the code in QQuickFlickable::wheelEvent)

And this issue became more obvious after ec4368adb7a32824f3ea80045198e78995fc82dc.
Since [vMoved] can now be reset to false in all boundsBehavior.

So we need to also check [vData.dragging] to make sure we are correctly in the first move.

Pick-to: 6.9 6.10
Fixes: QTBUG-137323
Change-Id: Ie2222e42c6628df62221ae8f63a55e7bf70d80f0
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>