qmlls: add internal --inputFile option
Add an option to qmlls to read from a file instead of stdin. This allows
to run qmlls in the macOs profiler. Make the option internal as normal
users of qmlls wouldn't need to read jsonrpc commands from a file.

Also remove void tst_qmlls_cli::cleanup() that was redundant from the
test (it does the same as void tst_qmlls_cli::stopServerImpl()), and
move the disconnect call to stopServerImpl() to merge both scopeguards
in the warnings() method.

Change-Id: Iff31ab33ffdcb2ac7237b2e239c7db06d11de81a
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>