qmlformat: make AstComments::ensureCommentForNode private
The method returns a reference that is invalidated on further insertion
of comments, like calling ensureCommentForNode() multiple time and using
the returned values for example. This might lead to more memory errors
like QTBUG-137413.

Make the method private to avoid further misuses.

Also add a comment on commentForNode() about the validity of the returned
pointer.

Task-number: QTBUG-137413
Change-Id: Ic785de96b8e8ff2ed108404c20b57296a91fdb93
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>