QtQml: Use QQmlScriptBlob in Qt.include
Invent a new URL fragment "include" for this. We need to make sure the
included JS code inherits the context.

We want to get rid of the local file loading in qv4engine.cpp since it
undermines network transparency. Qt.include therefore has to use the
type loader. For remote files it so far used a hand-rolled network
loading machinery which duplicated the type loader code. By using the
type loader right away, this becomes unnecessary.

Change-Id: Iae5c1ad764b98b101c1d90dbb78e46d3950541aa
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>