qmlls: move HelpEngine into QQmlCodeModel
With QTBUG-134308 we might need to support multiple documentation roots,
so move the HelpEngine into QQmlCodeModel. It also seems to be a better
place for the HelpEngine as it avoids connecting a signal to update the
HelpEngine on a documentationRootPath change.

This is a preparation step for QTBUG-134308 that aims to allow multiple
QQmlCodeModels in qmlls for the multi-workspace setup.

Task-number: QTBUG-134308
Change-Id: I7d6a1c1318f89309b13f7dd506e5d5f21fb6bdee
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>