qmlls: swap QQmlCodeModel with QQmlCodeModelManager
Replace all usages of QQmlCodeModel with the newly introduced
QQmlCodeModelManager, and add qmlls tests for the WS feature.

Adapt the calls to m_codeModel->registeredTokens() and helpManager
to pass the uri so the manager can call the method on the correct
codemodel.

Add tests where qmlls is run on multiple workspaces.

Fixes: QTBUG-134308
Change-Id: I6ba918bd37f78f892bfaead86a0e216cbca7d7f1
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>