Positioners: hide zero-width or height items
The documentation for e.g. Row says:

    If an item within a Row is not visible, or if it has a width or
    height of 0, the item will not be laid out and it will not be
    visible within the row.

Before this patch, the item wouldn't actually be hidden if its width
or height was zero, but just moved into a list of unpositioned items
and treated as if it was.

This patch fixes that by culling the item.

Fix and add rows to tst_qquickcontainer::zeroSize, which would
otherwise fail.

Fixes: QTBUG-132607
Pick-to: 6.8 6.9 6.10
Change-Id: I04c8a3f1ffb9e4a649024087626c4d7846ca01a0
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>